[UHYU-155] feat: home api 관련 타입 파일 분리
