+++ src/features/home/hooks/useBenefitsQuery.ts
@@ -1,8 +1,9 @@
+import type { Benefit } from '@home/api/home.types';
 import { useQuery } from '@tanstack/react-query';
 import { fetchBenefits } from '../api/homeApi';
 
 export const useBenefitsQuery = (grade: string) => {
-  return useQuery({
+  return useQuery<Benefit[]>({
     queryKey: ['benefits', grade],
     queryFn: () => fetchBenefits(grade),
     enabled: !!grade, // grade가 있어야 요청함
