+++ src/features/home/hooks/useNearbyStoresQuery.ts
@@ -1,9 +1,10 @@
 // src/features/home/hooks/useNearbyStoresQuery.ts
+import type { NearbyStore } from '@home/api/home.types';
 import { useQuery } from '@tanstack/react-query';
 import { fetchNearbyStores } from '../api/homeApi';
 
 export const useNearbyStoresQuery = () => {
-  return useQuery({
+  return useQuery<NearbyStore[]>({
     queryKey: ['nearbyStores'],
     queryFn: fetchNearbyStores,
   });
