+++ src/features/home/hooks/useUserInfo.ts
@@ -1,6 +1,6 @@
 // src/features/user/hooks/useUserInfoQuery.ts
-import type { UserInfo } from '@/features/home/api/homeApi';
 import { fetchUserInfo } from '@/features/home/api/homeApi';
+import type { UserInfo } from '@home/api/home.types';
 import { useQuery } from '@tanstack/react-query';
 
 export const useUserInfoQuery = () =>
